refactor: api서버 코드 정리
