+++ Backend/src/main/java/com/pg/programmerground/model/MemberGithubInfoRepository.java
@@ -1,9 +1,8 @@
 package com.pg.programmerground.model;
 
 
-import com.pg.programmerground.domain.MemberGithubInfo;
+import com.pg.programmerground.domain.github.MemberGithubInfo;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface MemberGithubInfoRepository extends JpaRepository<MemberGithubInfo, Long> {
-
 }
